requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
APScheduler==3.10.4
requests==2.31.0
Werkzeug==2.3.7
python-dotenv==1.0.0
ldap3==2.9.1
